Refactor: 도메인 레포 위치 변경 및 jpa리포지토리 도입
