+++ src/main/java/org/poolc/api/activity/repository/TagRepository.java
@@ -1,0 +1,8 @@
+package org.poolc.api.activity.repository;
+
+import org.poolc.api.activity.domain.Tag;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TagRepository extends JpaRepository<Tag, Long> {
+
+}
